+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add performance counters to producer ([#2424](https://github.com/infinyon/fluvio/issues/2424))
 * Upgrade to fluvio-future 0.4.0 ([#2470](https://github.com/infinyon/fluvio/pull/2470))
 * Add support to detecting smartmodule type from WASM payload on SPU  ([#2457](https://github.com/infinyon/fluvio/issues/2457))
+* Require `version` field in connector yaml. ([#2472](https://github.com/infinyon/fluvio/pull/2472))
 
 ## Platform Version 0.9.30 - 2022-06-29
 * Improve CLI error output when log_dir isn't writable ([#2425](https://github.com/infinyon/fluvio/pull/2425))